+++ CHANGELOG.md
@@ -5,6 +5,6 @@
 ### Features
 
 - Matchers to validate existing file and directory
-- Matcher for readable file
+- Matchers for file accessibility: read / write / execute
 
 [Unreleased]: https://github.com/clean-code-rocks/hamcrest-java-file/commits/main